(PUP-1526) Add environment settings for package type
This patch adds to the package type the `:settable_environment` feature
and the `:environment` parameter used by that feature.  `:environment`
is either a string of the form "NAME=VALUE" or an array of such strings.

The Puppet::Provider::Package is modified to include wrapped versions
of the `execute`, `execpipe`, and `has_command` methods which include
the custom environment if the calling resource defines it.  This
includes a required modification to the `CommandDefiner` class in
Puppet::Provider to allow the executor and resolver to be overridden.

All package providers are modified to use the Puppet::Provider::Package
versions of `execute`, `execpipe`, and `has_command`.

The `yum` package provider is updated to actually support the
`environment` parameter and `:settable_environment` feature.

Finally, the `spec` tests for the `package` type and `yum` provider
are updated to test for the `:settable_environment` feature and the
`environment` parameter.