(PUP-6137) Add ability to name Object type
Before this commit, all Object type declarations were anonymous. This
commit changes that so that:

a) An Object can have a name.
b) An alias of an anonymous object is instead promoted to a named object
c) An alias of a named object will have precedence (override).
d) Only objects added using an alias construct are known to the loader