(PUP-6102) Add support to handle 'shadowing'
Before this, the only way to check what was already loaded
was to load an entity. This caused problems when checking if something
about to be added should be added (as this recursively would
cause the very thing again being loaded)

Now, a new mechanism is added to the loaders where it is possible to
lookup an already loaded entity (will not trigger loading if not already
loaded).

A mechanism is also added that allows a loader to define if shadowing is
allowed or not (it is allowed for the predefined loader used in 4.x
functions; or it would be exposed to resource types in modules). It is
also allowed in the 3.x resource loader as it loads 3.x resources into a
loader that is below the static loader where aliases are already bound.