(maint) Use TypeFormatter.capitalize segmetns instead of just capitalize
Before this commit, the Runtime3TypeLoader would simply capitalize the
type name before creating a Resource type. This used to work since the
resource type would downcase the name anyway, but now the name is
checked against CLASS_REF_EXT which doesn't allow a qualified name with
segments that are not capitalized.