(PUP-5929) Add type mapping capability to parser and validator
This commit adds the ability to declare a mapping between a runtime
type pattern and a puppet type pattern to the PL parser and validator.
A mapping can be declares as:
```
type Runtime(ruby, [<pattern>, <repl>]) = [<pattern>, <repl>]
```
where <pattern> are regexps and <repl> are strings that references
groups in the regexps using \<n> syntax.