(PUP-6168) Fix problem with Runtime3Loader double updates
The Runtime3Loader calls on the known_resource_types of the current
environment to find a type. That call might ultimately result in that
the loader is being updated with the queried entry. When the call
returns, the loader must therefore check if it has the entry before
it tries to store it. This commit adds that check.