(PUP-6167) Fix problem default Tuple having max size of zero
A default tuple has no size, but when compared it creates a range based
on the size of it's contained types. The default tuple has an empty list
of types which resulted in the range 0,0.

This commit changes this so that the computed range is 0 - infinity when
there are no types.