(PUP-5105) Add ci:test:quick for per-commit reduced subset of tests
This commit adds a new acceptance rake task, `ci:test:quick`,
which runs a reduced subset of tests. The task is intended to be
run on a per-commit basis rather than the full suite, in order to
reduce time required for the development feedback cycle.

This task is *not* intended to replace puppet-agent's full suite
pipeline, which will still run all tests nightly.

The task includes a small set of tests which are an attempt at a
representative sample of the full suite. In addition, any tests
modified within the last two weeks are automatically added to the
list and run as well.

[skip ci]