(PUP-5353) Don't try to disable static systemd services
Static systemd services are defined as services that cannot
be enabled or disabled and are usually dependencies or tasks
of other services  Previously, Puppet would erroneously attempt
to disable these services, which is a no-op on the system. Upon
doing so, Puppet was fooled into believing that the service was
being successfully disabled in every run.

This commit updates the systemd provider to not attempt to disable
static services, and log a debugging message when a manifest
attempts to.

In addition, a new 'cached_enabled?' method is added which prevents
the provider from unnecessarily shelling out to systemctl multiple
times.

Thanks to Michael Smith for the research and part of the implementation
for this fix!