(PUP-6176) Fix error when a resource reference is appointing undef
PUP-6028 fixed a bug in how the resource type and title was extracted
where the case when both where `nil` was undetected. This revealed
another bug in the `RelationshipValidator` where it attempts to create
a resource from a parameter with the value `nil`.

This commit ensures that no attempt is made to create a resource from
`nil` during parameter validation.