(PUP-2354) Do not retrieve client cert when generating new cert
When using `allow_duplicate_certs = true` and autosign mode on master
server, we expect that the client does not retrieve the previous signed
cert present on the master. This fix disable the search mechanism when
submitting the csr to let the server sign it, and let the agent download
the freshly signed cert instead of the previous one. It only work on the
first run with autosign enable obviously.

This is needed because we are using autoscaling group and preemptible
instances on our cloud provider. In this case the name and the certname
of the new instance is the same of the previous one.