(maint) Remove TypeReference string/name[param] ambiguity
Some code relied on that the name of a TypeReference actually was a
type in String form whereas other code assumed that a TypeReference
would keep the name and parameters in separate objects. This commit
simplifies that TypeReference so that it only contains the string
that represents the type that it is referring to.