(PUP-3740) Require at least ruby 1.9.3 for the puppet gem
Puppet gems do not require at least ruby 1.9.3, even though the code
will not run on earlier versions.

Add the ruby and rubygems versions to project_data.yaml, which will be
copied by the `rake package:gem` task when building the gem.

I copied the values from the .gemspec, which is used when another
project, e.g. module, depends on a source checkout of puppet. The values
in .gemspec and ext/project_data.yaml should be kept insync.

[skip ci]