(PUP-6046) Add a 'dig' function that allows user to dig into a structure
The dig function allows "digging into" a structure by specifying a
sequence of keys/indexes into a structure of Hash/Array values.

An `undef` value is returned if a key does not exists.
If a key is applied to a a result that is not an Array or Hash (but not
Undef) an error is raised.