(PUP-6046) Add a 'then' function for convenient computation of sequence
This adds a 'then' function which is useful when needing a chain of
computations where intermediate results may be `undef`.

$x.then |$x| { $x[a] }.then |$x| { $x * 2 }

The function returns `undef` if given `undef` and otherwise the result
of yielding the argument to the given block.

Thus, an intermediate `undef` result anywhere in the chain ripples 
through the chain without causing any of the blocks to be called.

The end result can then be asserted with assert_type (if an `undef` 
result is not acceptable, or if a default value should be used as the
final result.