(PUP-6046) Add the 'lest' function (if_undef)
This adds the `lest` function which yields to a lambda (without
arguments) if the argument given to `lest` is `undef`. The value of
the lambda is returned. If the function is given a value other than
`undef` this value is returned, and the lambda is not called.

The `lest` function is part of the family of functions that include
`dig`, `then`, and `with` (and `assert_type`) with which it is easy to
construct a chain of operations without the need of conditionals and
extra variables.

The `lest` function is the reverse of `then` (`then` passes on `undef`
and `lest` passes on values matching `NotUndef`).