(PUP-6230) Fix correct return of undef values found in yaml
Before this commit, the Hiera Data Provider for Yaml would incorrectly
translate a `nil` value to a `hash`. While this is correct at the
very top (Hiera returns nil for an empty file), it is not correct when
a key is found and its corresponding value is `nil`. This commit ensures
that the conversion to a hash only happens at the top.