(PUP-6224) Improve TypeFormatter so that it can "pretty print"
This commit adds rudimentary "pretty print" support to the type
formatter in the form of a new method TypeFormatter#indented_print.
The method works like print but takes two additional arguments,
`indent_level` and `indent_width` that controls the actual indentation.

The implementation provided here will put all hash entries on a new
indented line and increase the indent level with one for each nested
hash that is formatted.