(PUP-6224) Add a Ruby generator that emits code based on Object types
This commit adds a Ruby generator capable of generating a default
implementation of an Object type. Instances of the object can be
created either from a hash (validated by the #i12n_type of the
Object) or by a constructor using anonymous parameters.

The generator is capable of generating both anonymous ruby classes
and a static implementation intended to be written to files and
then included in a code base.