Provide custom execution functions for package providers
The `package` provider is modified to provide custom `execute`,
`execpipe`, and `has_command` functions.  These functions
allow modification of command execution based on package
features.  Currently, the only modified command execution is
for the `:settable_environment` feature, which modifies the
`custom_environment` of the command execution.

The `yum` provider's custom execute function is removed in
favor of the parent package provider's equivalent.