(PUP-6225) Fix RubyGenerator so that it emits a 'raw' hash constructor
When a dispatcher is used to select a method, it will assert that the
arguments conform with the parameter types declared in a dispatch
declaration. The generated `#from_hash` method was therefore divided
in two where the first method will assert the hash type and then
delegate to the second `#from_asserted_hash`. A dispatcher may then also
use the second method to avoid that the already asserted hash is
checked again.