(PUP-6225) Add 'new_function' to the Object type
Adds the 'new_function' to the Object type. The function will check
if an implementation class exists for the type, and if so, that class
will be responsible for creating new instances. If no mapping exists,
then the RubyGenerator will be asked to generate a default
implementation in the form of an anonymous class.