(PUP-6233) Serivce SMF restart should use -s
The restartcmd in smf.rb needs to add the -s flag to make
sure that service restarts are synchronous. Without this
patch smf will return prior to the completion of the restart.

(PUP-6238) Serivce SMF should set has_feature :refreshable

SMF supports the concept of restart instead of calling start/stop
on each service. This patch informs the higher level service
provider of that capability.