(PUP-6235) Change PMT tests to examples in build test
The PMT skeleton has been changed to name the 'tests' directory
to 'examples' in commit 420cb9f6bb. This caused the
'build_should_not_allow_symlinks' test to fail because the test
was using the 'tests' directory.

This commit renames the 'tests' directory to 'examples' in the
'build_should_not_allow_symlinks' test.

[skip-ci]