(maint) Remove dead code in Evaluator#is_match?
The method Evaluator#is_match? had an if-statement with ignored result.
Or almost ignored. It would cause the method to return `false` when
matching a `Regexp` against something that wasn't a `String`. That test
is however also present in the `CompareOperator#match` that is called
later. This commit removes the if-statement.