(PUP-6221) Add Semantic versioning to the Puppet Type System
This commit adds the two types SemVer and SemVerRange to the Puppet
Type System. The SemVer is a parameterized type and it's arguments can
be either strings or SemVerRange instances.

A SemVer type is constrained by it's contained SemVerRanges. If it has
no ranges, then the type represents all SemVers. If it has ranges, then
the type represents all types that matches any of the ranges.

The ranges of a SemVer type will be normalized so that any range that
intersecs with another range automatically forms a new range. This is
needed to get the type arithmetic working correctly in cases like:

$x = SemVer['>=1.0.0 <2.0.0', '>=1.5.0 <3.0.0']
$y = SemVer['>=1.2.0 <2.8.0']
$x > $y == true