(PUP-6221) Use type aliases and manual assertion of strings
This commit removes some complex regular expressions that were present
in the string dispatcher for SemVer and SemVerRange in favor of just
calling the respective #parse function which gives a better error.

It also provides local aliases for the involved types to improve the
user experience when faulty arguments are used.