(PUP-6178) Look for hiera.yaml in codedir first, then confdir
This commit ensures that when Puppet is initialized, it will first check
if a Hiera config file is specified using the `hiera_config` setting. If
it is, then that setting will be used (and an warning is issued if it
doesn't exist). When`hiera_config` is not set, Puppet will first look
`hiera.yaml` in the codedir and if no such file exists, it will default
to `hiera.yaml` in confdir.