(PUP-6046) Make dig return undef if a key is undef
Before this, if user entered something that evaluated to a call
on the form:
```
$data.dig(key, undef, key)
```
The undef key could potentially be used in an attempt to access
an index in an array and this would fail with an error. (In a hash it
would possibly return a value, if there was a value for the key undef).

Since use of undef keys in hashes are discouraged the function will now
stop and return undef if either an intermediate value, or a key is
undef.