(PUP-3827) Throw proper HTTP errors from indirected routes
Previously we threw ArgumentErrors when we received a malformed request
for the IndirectedRoutes class. This conflated an error in calling a
Ruby method and an error by the consumer calling an api endpoint, and
consequently always returned 400 statuses for internal errors.

As part of PUP-3827, higher level middleware now handles the response
formatting for errors. The higher level middleware, however, makes the
distinction between ArgumentErrors to Ruby functions and BadRequests to
the API endpoint (and expects throwers to as well).

Since IndirectedRoutes did not make this distinction all invalid
environment requests were then being treated as ServerErrors, and
returning 500 statuses.

This patch updates IndirectedRoutes to throw proper HTTPErrors instead
of ArguemntErrors when a client has submitted a BadRequest, used a
MethodNotAllowed, or the environment is NotFound.