(PUP-6301) Make AIX service test setup idempotent
This commit alters the setup step for ensuring that the test service is
stopped. It removes the validation that the service changed state. This allows
the step to pass if the service was already stopped prior to running the
setup step that stops it.

Prior to this commit, if the system did not automatically start the service
upon creation the test would fail during the setup step intended to ensure
that it was stopped.