(maint) Remove code that corrupts regexp patterns
Before this commit, the TypeFactory#regexp method would always
convert a Pattern to a String by doing inspect and then remove the
first and last character (i.e. strip of the slashes). This breaks
regexps that contains options.

Although such regexps are not allowed in the Puppet Language, it's
still of value to be able to use them internally as the URI#regexp for
example, uses the 'x' option.