(PUP-6220) Add TypeSet type to the Puppet Type system
This commit adds the TypeSet type to the Puppet Type system parser,
evaluator, and string formatter.

A TypeSet can define new types using a Hash construct where the key
is the name of the object and the value is a Type. The Type is typically
an Object[] declaration but can also be something else, in which case
the key => value pair is treated as a type alias.

A TypeSet may also contain references to other TypeSets. Just references
are resolved during the normal #resolve phase for the type (called from
the TypeParser when the type is loaded). The references are resolved
using a loader.

Types in the TypeSet itself are resolved using a special loader that
is aware of the TypeSet so that all types can be cross-referenced. This
also means that types within the TypeSet can access types from the
referenced TypeSets.