(PUP-3964) acceptance: ensure provider from same env as custom type
This change adds a test to ensure a custom type and provider both come
from the same/correct environment as the rest of the agent run. We
create two modules with the same type and provider, one in production
and one in another environment. Only the file produced by the
type/provider in the 'other environment' shall be ensured.

[skip ci]