(PUP-6320) Fix error when reporting zero sized array mismatch
Fixes a bug in the TypeMismatchDescriber that caused the type mismatch
reported when passing a zero sized array where a non empty array is
required to report an element type mismatch rather than the size
mismatch.

Also fixes so that array mismatches are reported with correct index
rather than just stating that there's a mismatch between an Array
and a Tuple where some element in the Tuple doesn't match the element
type of the Array.