(PUP-6336) Fix problem with diff Jruby/MRI in String#==
Puppet has a monkey patch for Symbol#<=>(String) that enables
comparison. This is deprecated and slated for removal in puppet 5.0.0.

On JRuby, String#==(other) is implemented as `other<=>(self) == 0` which
causes the monkey patch in puppet to trigger.

In 1547756ad7caa5cb1f0add8161e51ac07f61bd67 we reversed a test from `x
== :undef` to `:undef == x`. And this caused the problem reported in
PUP-6336 as this logic now (via the MRI incompatible implementation in
JRuby) caused the monkey patch in puppet to kick in.

This commit changes the test `:undef == x` to `:undef.equal?(x)` which
avoids bothe the JRuby incompatibility, and the problem the change
solved in the first place (ending up in the same insane monkey patch).