(PUP-6339) Bind lookup adapter to compiler instead of environment
Before this commit, the LookupAdapter used for providing environment and
module data providers, was attached to the current environment. This
caused all caches to have the same lifecycle as the environment which
is wrong. The caches should not survive between compilations.

This commit changes the LookupAdapter so that it instead attaches itself
to the compiler. This ensures that a new LookupAdapter instance (and
and hence, new caches) are created for each compilation.