(doc) Recommend Puppet Extension add module name
Even with using the namespace ('user') as part of the PuppetX naming
convention, there still could be conflicts with other existing
modules in the same namespace. This could cause two files from
different modules that share the same name to overwrite each other
when they are synced to the same folder. It may not be immediately
obvious this is the reason when this occurs, and the results can
be disastrous. Due to these reasons, we should recommended that
Puppet Extensions submodules further define folder structure to
include the module name as well.