(PUP-3964) fix previous test commit teardown.
This test was previously not cleaning out its production site.pp, which
broke subsequent tests.  This test really should be using a temp
environmentpath, to be even safer, but this is a bit slower as it
requires a server restart.  This change also only cleans out the files
on the master, saving a bit of time.