PUP-6366 trusted facts with apache/nginx and puppetmaster unicorn
running unicorn behind nginx or apache reverse proxy, the
$trusted facts are not available to puppet.

The supported the +ExportCertData from Apache, only works
with Passenger module, but not with thin rack behind reverse
proxy, and esp. not with nginx.

Therefore I added an additionalheader that gets passed to unicorn: X-SSL-Client-Cert.
However, that header is sent as single line from Apache to unicorn,
and not as valid PEM encoded certificate. Therefore the gsub!
manipulations to restore a valid PEM certificate again.

To make use of it with Apache add this to the vhost:
RequestHeader set X-SSL-Client-Cert %{SSL_CLIENT_CERT}e

With nginx, there is a bit more trouble. Nginx has $ssl_client_cert
variable aswell, but nginx passes that variable on as multi-line header. Doh!
Unicorn doesn't like that at all.

To make it work with nginx, you need lua in nginx, then you need
something like this in your vhost:

 location / {
   set_by_lua $client_cert "if ngx.var.ssl_client_raw_cert then return ngx.var.ssl_client_raw_cert:gsub('\\n',' ') end";
   proxy_set_header X-SSL-Client-Cert $client_cert;
 }