(PUP-5831) Add a find_file function
When designing (a yet to be committed) binary_file function the ability
found in the file function to search for a file to read was better
implemented as a single general purpose function letting the binary file
function operate on a single file.

The added `find_file` searches for the first existing file among a given
set of paths; as individual arguments or in an array. The first absolute
or module relative path to an existing file is returned.

By separating this behavior into a separate file it is also possible for
the author to write conditional logic instead of just seeing a failure
from the `binary_file` read if none of the paths existed.