(pup-6380) Fix HTTP file source for GET-only URIs
The Problem:
When using an HTTP URI as the source of a `file` type, URIs that only
work for GET requests fail.

The Use Case:
Pre-signed Amazon S3 URIs include the HTTP method in the signature,
which is usually GET for downloading files. This makes the URI return a
403 when responding to a HEAD request, but a 200 when responding to a
GET. Github uses Amazon S3 to store project 'releases', snapshots of a
repository at a certain release, usually prepackaged for different
operating systems and architectures.

The Code:
The `file` type fires off a HEAD request to the file's source when
trying to get metadata for the file,
https://github.com/puppetlabs/puppet/blob/master/lib/puppet/type/file/source.rb#L188.
Through the indirectors, this eventually makes its way to
Puppet::Util::HttpProxy, when a method of `:head`.

The Changes:
There are two main changes here, the first being that we now pass both
the `uri.path` and `uri.query` to HTTP requests. URI query parameters
can contain necessary authorization tokens or signatures, and many URIs
will fail without them. Once support for Ruby 1.9 is dropped, the entire
URI object can be passed into HTTP requests, rather than needed to
reconstruct the URI.

The second change is when a HEAD request fails, attempt a GET request
to see if that works. This should only happen in the file metadata case,
as this is the only case I could find that specifically uses the
HttpProxy class to make a HEAD request.

If there's a better way to do this/implement this or it's already being
taken care of, please let me know, as I'm still pretty new to this repo :)