(PUP-5831) Improve error message from failed type assertion in new
Before this, there was no indication that the error message 
arising from the final type assertion performed by the function new was
related to the assertion. Instead, it could be read as new was called
with the wrong arguments.

Now the message says that conversion resulted in a non type compliant
instance in a clearer way (including mentioning the expected type).