(PUP-5831) Add base64 syntax checker to heredoc
This adds a base64 syntax checker to heredoc such that
@(END:base64)
  <base64text>
END

must be base64 compliant. When checking, whitespace is ignored.
This means that the check is somewhere between relaxed base64 encoding
which simply skips all characters outside valid range, and strict base64
that does not allow any whitespace.

A result that is valid with this checker is valid strict base64 if all
whitespace matching /\r?\n|[[:blank:]]/ is removed.