(PUP-6150) Don't include IP address in auth error response
Previously, an unauthenticated client could make a request to a route
protected by auth.conf and get back an error message including the IP address
and rule description. This is more information than is probably good to
provide over http, especially since it could potentially include internal IPs
of proxy servers, depending on how those are configured. This commit makes it
so that the response includes minimal information.