(PUP-3768) Scheduled_task keeps reapplying
Initially, puppet would re-apply a scheduled task even though
there was no change.  The reason being, only the first 256
characters of the arguments parameter (or rather, any
parameter that accepts a string as input) were being read in
for comparison.  Therefore, the first 256 characters of the
string were being compared to the full string, resulting in
changed status.

This commit increases the comparison limit for the command
and arguments parameters to 260 and 4096, respectively.
The 260 character limit is due to Windows API MAX_PATH
limit for pathnames.  The arguments parameter doesn't appear
to have an upper bound, so 4096 was chosen as a reasonably
large limit.  An integration test was included and the
should_modify acceptance test was changed, both to exercise
scheduled_task with larger string params.