(PUP-6394) Relax code id requirement when finding capabilities
Previously when looking up capability resources, the compiler would
lookup resources resources strictly by code id. This effectively meant
that a node couldn't run against a certain code version until all of its
dependencies had run against that version.

For iterative development on changes that don't actually touch those
dependencies, this is an extremely onerous requirement. Further, it's
not sufficient for a node to only run after its dependencies; if the
capabilities it requires are changing, it needs to be run *quickly*
after its dependencies, which is a property that isn't satisfied by the
ordering requirement alone.

In order to provide more flexibility around running agents, particularly
when capabilities aren't being affected by a given change, we now use
any available version of a consumed capability resource, rather than
requiring it come from the current code id.

This creates the possibility that a capability may be moved from one
node to another node, causing two copies to exist in the database for a
time. In that case we repeat the lookup, this time using code id to
disambiguate. That disambiguation prevents the catalog from referring to
the wrong version of a capability resource, though it does reintroduce
the requirement that the producer has run against the current code id.