(PUP-6268) New flags for `puppet cert` interactivity and formatting
In adding support for certificate extensions, we need to change the way
we display these cert information. While doing so the CLI leads have
asked that we update the `puppet cert sign` command as well to include
targeted usability improvements.

This commit plumbs the options for these changes through the `puppet
cert` application. It adds a `-h`, or `--human-readable` flag for pretty
printing cert information, a complementary `-m` or `--machine-readable`
flag for line-wise sed-able processing. It also adds a `--interactive`
or `-i` flag for triggering interactive behavior when signing certs and
a `--assume-yes` or `-y` to forcefully skip interactive behavior.