(PUP-6268) Update `puppet cert list` output for extensions
When dealing with custom extensions in CSRs and certs we need a way to
display that information to our users. Unfortunately, our current output
of `puppet cert list` provides a line-wise format that wraps and becomes
hard to read on smaller terminals, leaving little room to add this
information without larger changes.

This commit adds a small notation to the existing ("legacy") view to
denote that extension information is available for a cert or CSR.

In this commit we add a machine focused output that is also line-wise
orientated and includes custom extension information and expiration
timestamp.

We also provide a human focused, pretty printed, version of the full
information.

If a user uses the verbose flag without specifying the output type they
will see the full machine focused output since it is closest to the
current, line-wise, default legacy format.