(PUP-6376) Add failover logic to the agent
This adds master failover logic to the agent using the new
`server_list` setting. The failover happens early in the configurer,
with gating to ensure it doesn't happen when the legacy `server`
setting is being used, or if the configurer is actually an `apply`
run.

The use of the selected master is handled using Puppet's `lookup`
system. This allows us to scope the bound server info to all logic
that happens within a configuration run, without needing to refactor
the entire indirector.

The changes on the indirector side are fairly simple - check if
server/port have been bound with `lookup` and if not fall back to the
`server` and `masterport` settings.